backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiomysql
redis
python-nmap
requests
httpx
pydantic
